rename files to comply with the approach used in the project
