+++ components/wysiwyg/widget-block-edition/actions.js
