+++ components/wysiwyg/widget-block-edition/component.js
