+++ components/wysiwyg/widget-block-edition/default-state.js
