+++ components/wysiwyg/widget-block/actions.js
