+++ components/wysiwyg/widget-block/component.js
