+++ components/wysiwyg/widget-block/default-state.js
